requirements.txt
websockets==12.0
python-dotenv==1.0.0
flask==3.0.0
flask-socketio==5.3.5
python-socketio==5.10.0
